Move curves to math
